requirements.txt
PyQt5
Pillow
pynput
mss